quell %lu vs size_t warning
svn:revision:37726
svn:branch:trunk
svn:account:erikgreenwald

+++ src/librt/primitives/poly/poly.c
@@ -838,7 +838,7 @@ rt_pg_describe(struct bu_vls *str, const struct rt_db_internal *ip, int verbose,
 	fastf_t *n = pgp->poly[i].norms;
 
 	sprintf(buf, "\tPolygon %lu: (%d pts)\n",
-		i, pgp->poly[i].npts);
+		(long unsigned int)i, pgp->poly[i].npts);
 	bu_vls_strcat(str, buf);
 	for (j=0; j < pgp->poly[i].npts; j++) {
 	    sprintf(buf, "\t\tV (%g, %g, %g)\n\t\t N (%g, %g, %g)\n",
